chore(segment): Move package segment out of install
