+++ internal/install/execution/segment_report_test.go
@@ -8,8 +8,8 @@ import (
 	log "github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/require"
 
-	"github.com/newrelic/newrelic-cli/internal/install/segment"
 	"github.com/newrelic/newrelic-cli/internal/install/types"
+	"github.com/newrelic/newrelic-cli/internal/segment"
 )
 
 func TestSegmentReporter_InstallStartedShouldHaveNoError(t *testing.T) {
