+++ internal/segment/client.go
