+++ internal/segment/client_test.go
